+++ vite-project/src/components/Header/Header.jsx
@@ -1,3 +1,5 @@
+import PopUser from "../Popups/PopUser/PopUser";
+
 function Header() {
     return <header className="header">
     <div className="container">
@@ -11,16 +13,7 @@ function Header() {
             <nav className="header__nav">
                 <button className="header__btn-main-new _hover01" id="btnMainNew"><a href="#popNewCard">Создать новую задачу</a></button>
                 <a href="#user-set-target" className="header__user _hover02">Ivan Ivanov</a>
-                <div className="header__pop-user-set pop-user-set" id="user-set-target">
-                    <a href="">x</a> 
-                    <p className="pop-user-set__name">Ivan Ivanov</p>
-                    <p className="pop-user-set__mail">ivan.ivanov@gmail.com</p>
-                    <div className="pop-user-set__theme">
-                        <p>Темная тема</p>
-                        <input type="checkbox" className="checkbox" name="checkbox"/>
-                    </div>
-                    <button type="button" className="_hover03"><a href="#popExit">Выйти</a></button>
-                </div>
+                <PopUser />
             </nav>					
         </div>
     </div>
